progression model created, mysql schema created
