+++ src/main/java/com/greenfox/tribes1/Progression/PBuilding.java
@@ -1,0 +1,9 @@
+package com.greenfox.tribes1.Progression;
+
+import javax.persistence.DiscriminatorValue;
+import javax.persistence.Entity;
+
+@Entity
+@DiscriminatorValue("PBuilding")
+public class PBuilding extends Progression {
+}
