+++ src/main/java/com/greenfox/tribes1/Progression/PTroop.java
@@ -1,0 +1,10 @@
+package com.greenfox.tribes1.Progression;
+
+import javax.persistence.DiscriminatorValue;
+import javax.persistence.Entity;
+
+@Entity
+@DiscriminatorValue("PTroop")
+public class PTroop extends Progression{
+
+}
