--- src/main/java/com/greenfox/tribes1/Upgrade/Progression.java
@@ -1,31 +1,0 @@
-package com.greenfox.tribes1.Upgrade;
-
-import com.greenfox.tribes1.Kingdom.Kingdom;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.sql.Timestamp;
-
-@Entity
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public abstract class Progression {
-  private Long id;
-  private boolean toCreate;
-  private String type;
-  private Long model_id;
-  private Timestamp started_at;
-
-  @ManyToOne(cascade = CascadeType.PERSIST, fetch = FetchType.EAGER)
-  @JoinTable(name = "kingdom_progression",
-          joinColumns = @JoinColumn(name = "progression_id", referencedColumnName = "id"),
-          inverseJoinColumns = @JoinColumn(name = "kingdom_id", referencedColumnName = "id"))
-  private Kingdom kingdom;
-
-
-}
