--- src/main/java/com/greenfox/tribes1/Upgrade/Progression_Building.java
@@ -1,5 +1,0 @@
-package com.greenfox.tribes1.Upgrade;
-
-
-public class Progression_Building extends Progression {
-}
