+++ src/main/resources/db/migration/mySQL/V1_5__init.sql
@@ -1,15 +1,18 @@
-CREATE TABLE `upgrade` (
-  `id`         bigint(20) NOT NULL AUTO_INCREMENT,
-  `is_created`   boolean,
-  `timestamp` timestamp(6),
+CREATE TABLE `progression` (
+  `id`         bigint(20) AUTO_INCREMENT,
+  `model_id`       bigint(20),
+  `is_create`   boolean,
+  `finished_at` timestamp(6),
+  `progression_type` varchar (20),
+
   PRIMARY KEY (`id`)
 );
 
-CREATE TABLE `upgrade_kingdom` (
+CREATE TABLE `kingdom_progression` (
   `kingdom_id`          bigint(20) not null,
-  `upgrade_id` bigint(20) NOT NULL,
-  PRIMARY KEY (`upgrade_id`),
-  KEY buildings_id (`upgrade_id`),
-  CONSTRAINT kingdom_upgrade_kingdom_id FOREIGN KEY (`kingdom_id`) REFERENCES `kingdom` (`id`),
-  CONSTRAINT kingdom_upgrade_ugprade_id FOREIGN KEY (`upgrade_id`) REFERENCES `upgrade` (`id`)
+  `progression_id` bigint(20) NOT NULL,
+  PRIMARY KEY (`progression_id`),
+  KEY progression_id (`progression_id`),
+  CONSTRAINT kingdom_progression_kingdom_id FOREIGN KEY (`kingdom_id`) REFERENCES `kingdom` (`id`),
+  CONSTRAINT kingdom_progression_progression_id FOREIGN KEY (`progression_id`) REFERENCES `progression` (`id`)
 );
